Update Raycaster.html
My graphics card started crying and locked itself in its room and didn't allow the desktop to be initialized on startup because it was overworked and underpaid. That's all fixed now, but long story short, I've toned down the scale on which the program simulates so that hopefully your graphics card won't be as strained now. The option is still there to increase or decrease the resolution; I've simply decreased it in the code.